requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
flask-cors>=4.0.0
gunicorn>=21.2.0
pyjwt>=2.8.0

